Optimized generation of visual mesh by combining surfaces when possible
Surfaces on a brush that share the same material will now be combined, reducing the number of draw calls required.  This has the bonus benefit of improving generation time as well.

Also made it possible to turn off generation of the visual mesh entirely, for things like Areas that would only use the collision shape.